requirements.txt
streamlit==1.28.1
pandas==2.1.2
numpy==1.26.1
scikit-learn==1.3.2
plotly==5.18.0
requests==2.31.0
beautifulsoup4==4.12.2
validators==0.22.0
tldextract==5.1.1
lime==0.2.0.1
textstat==0.7.3
textblob==0.17.1
pillow==10.1.0
pytesseract==0.3.10
wordcloud==1.9.2
matplotlib==3.8.1
python-magic==0.4.27
python-whois==0.8.0
nltk==3.8.1
